Initial fuzzing with multiple arguments for execFile, fork, and spawn
Add initial fuzz target for fuzzing the child_process functions
`execFile`, `fork`, and `spawn` with more than one argument. For
simplicity, the approach here is to split the buffer on newlines as
delimeters for the arguments. This is not ideal (because newlines are
important characters to fuzz test, per release v1.5.8) and so subject
to change.