Fuzz with multiple arguments for execFile, fork, and spawn (#342)
* Initial fuzzing with multiple arguments for execFile, fork, and spawn

Add initial fuzz target for fuzzing the child_process functions
`execFile`, `fork`, and `spawn` with more than one argument. For
simplicity, the approach here is to split the buffer on newlines as
delimeters for the arguments. This is not ideal (because newlines are
important characters to fuzz test, per release v1.5.8) and so subject
to change.

* Correct argument splitting

Omit second argument, which resulted in `::split` returning an empty
array.

* Merge try-catches for fork testing

Since separating them serves no purpose.

* Fix multi-argument fuzzing on Windows

Fix the multi-argument fuzz target on Windows by:

1. Not `disabl[ing]ExtraWindowsPreparations`, which are required in this
   case (based on experimental evidence).
2. Filtering out empty arguments for PowerShell, which appears to filter
   them out (as opposed to all other tested shells).