Set up continuous testing workflow
Set up a new workflow for continuous testing that continuously runs
"discovery" tests (like fuzz tests) on a schedule. The re-usable fuzz
workflow is used here, and with a longer duration as it matters less how
long it takes on a schedule.

Correspondingly, remove the schedule trigger from the pre-existing test
workflow (now only running on pushes and Pull Requests).