Fix null-character related false-positive in fuzz targets
Fix a mistake in (some of) the fuzz targets for PowerShell (execFile and
spawn). The problem was caused by lines containing only null characters,
which are treated equivalently to empty lines by PowerShell, but were
not stripped when obtaining the expected output (as seen on, e.g. line
88 of spawn.test.cjs). This has been fixed by filtering with null-char
cleared strings. An entry in the fuzz corpus was included to prevent
regression against this problem.