refactor: fix trivial regular expression linting errors
Fix all regular expression linting errors that are "trivial" - in
essence those that can be fixed automatically.

One rule was excluded from this, namely 'regexp/require-unicode-regexp',
because of the significant amount of changes (essentially all regular
expression in the project) combined with historical problems with the
`u` flag. Therefore, this rule will be evaluated separately.