fix: potentially quadratic runtime regexps for Bash escaping
Fix the warning by 'eslint-plugin-regexp' [1] about a potentially
exploitable quadratic runtime regular expression in the source code
(line 55):

> Any attack string /\{+/ plus some rejecting suffix will cause
> quadratic runtime because of this quantifier

and (line 56):

>  Any attack string /:+/ plus some rejecting suffix will cause
>  quadratic runtime because of this quantifier

by simplifying the regular expressions. As a result, the respective
characters will be escaped in cases where it's not necessary. This is
not a problem because the string received by the program will still be
correct. This is beneficial as it avoids the quadratic runtime.

The only affected regular expression in the package code is for escaping
for bash with the `interpolation` option set to `true`.

The test fixtures for Bash were updated accordingly and improved upon
with new test cases for previously untested strings that should be
escaped that were identified while trying to adjust the corresponding
regular expressions to avoid quadratic runtime.

Similarly, a new entry was added to the fuzz corpus that was found after
the regular expression on line 55 was changed.

--
1. https://github.com/ota-meshi/eslint-plugin-regexp