chore: configure eslint-plugin-regexp to reporting potential problems
Update the ESLint configuration to eslint-plugin-regexp to report not
only certain violations but also potential violations. This was chosen
because currently there's no potential violations, indicating a
stability level of these rules that's compatible with this project. If
this turns out to be incorrect in the future, these changes may be
reverted.