chore: fix potentially quadratic runtime regexps for fuzzing
Fix the quadratic runtime regular expressions in the common fuzzing
logic by adding negative look-behinds for the target being matched.
This ensures that the match fails immediately at character `n+1` if
it failed at character `n` (because character `n` is in the character
range).

Also simplify the regular expression for PowerShell by consolidating all
the individual whitespace characters into the `\s` matcher. Note that
`\s` excludes `\u0085` so it MUST be included separately.