Lint regular expressions & fix problematic ones (#373)
Add ESLint [1] with the eslint-plugin-regexp plugin [2] to start linting
and enforcing styles for regular expressions. Given the heavy reliance
of this package on regular expression, this should improve the package
as well as remove any ambiguity on how regular expression _should_ be
written.

ESLint is configured explicitly to only lint with the "regexp" plugin
(i.e. eslint-plugin-regexp [2]), which is entirely manually configured
to ensure 1) that everything is enabled (^1) and 2) it's clear and
explicit what is being linted for.

`eslint` was added as an npm script, and is used in GitHub Actions as
well as the git hooks to continuously evaluate and report on regular
expressions.

First off, all trivial errors from eslint-plugin-regexp [2] have been
fixed by running `npm run eslint -- --fix`. This includes (but is not
limited to):

- Ordering characters and character groups.
- Simplifying expressions (e.g. use `\0` over `\u0000`).
- Using character classes over single-character alternatives (groups of
  2 single-character alternatives were converted manually).
- Ungrouping unused capturing groups.

One notable automatically fixable rule is regarding the use `u` flag on
all regular expressions (`require-unicode-regexp`). According to [2, 3]
this:

> Make[s] the regular expression [handle] UTF-16 surrogate pairs
> correctly. [And] make[s] the regular expression [throw] syntax
> errors early [by] disabling Annex B extensions.

This change was also tested with Stryker following the removal of this
flag in d4f4c3a due to a bug in Stryker 6.0.1 when used in Node.js 18.

Second, two regular expressions for Bash were updated following the
`regexp/no-super-linear-backtracking` and `regexp/no-super-linear-move`.

- `regexp/no-super-linear-backtracking`: This involved replacing the
  first `[^]*?` group on line 55 by `[^,.]` to avoid the possibility of
  exponential backtracking when exchanging characters with the second
  `[^]*?` (through `(?:\,|\.)` because both "." and "," are matched by
  `[^]`).
- `regexp/no-super-linear-move`: 
  - Line 55: This regular expression was made more explicit to prevent
    quadratic runtime by being more explicit about which characters are
    allowed when as well as adding a negative look-behind for curly
    braces prior to the first curly brace (to prevent needlessly
    retrying the expression if the first in a sequency of opening curly
    braces fails).
  - Line 56 (now 58): This regular expression was simplified to escape
    more often than is really necessary. But as unnecessarily escaped
    characters are still interpreted correctly, this is preferred over a
    super-linear expression. In particular, the `[^]*:` part of the RE
    must match for any number of `:` to properly escape, this inherently
    causes the expression to quadratic in runtime for successive colons.
  - The work on this also resulted in the new test cases in the Unix
    fixtures as well as the new fuzz corpus entry.

These rules also flagged some problems in the regular expression found
in the fuzzing logic (`_common.cjs`), these were fixed by adding
negative look-behind for the appropriate character classes (e.g.
`(?<!\\)`) as well as removing some unnecessary submatches (e.g. `\0*`).

The three fixes for super-linear regular expression have been added to
the CHANGELOG as these have a notable effect for end-users w.r.t.
performance (for certain classes of input strings).

--
1. https://eslint.org/
2. https://github.com/ota-meshi/eslint-plugin-regexp
3. https://eslint.org/docs/latest/rules/require-unicode-regexp

^1 The "regexp/prefer-named-capture-group" is disabled because it's
(currently) not desired, unnamed groups seem okay for now.