Create and use a test macro to test poisening effect
Create a new test macro that can be used to test the effects of
globals poisoning on a function. In particular, it runs the provided
function while observing calls to modified globals. If, after calling
the function, the modified global was called this is means the test
fails - and the function is affected by globals poisoning.

This may be a problem because the value of globals may change at
runtime. Hence the test. The test is currently limited to two globals
that are known to be used in this project. This and other gotcha's
about the macro can be found in its documentation.

The macro is already employed to test this project. Like the macro for
prototype polution, this is currently being tested in the integration
tests.