Ignore `process.{getuid,getgid}` for poisoning testing
These happen in a dependency. As, at this point in time, I have no
reason to believe the impact of poisoning in user land should be
considered a security problem [1], I don't think it's necessary to ask
the dependency to update. We could still decide to harden the current
package against poisoning where possible.