Update the Release Guidelines
Change the automated release guidelines to instruct the reader to follow
the guidelines in the Pull Request comment (establishing a single source
of truth).

Also take what used to be described about annotating the git tag from
automated releases section and put it in the manual releases section for
clarity.