Add script to easily run compatibility tests locally (#415)
* Create script to run full compatibility tests locally

Add a script that allows you to easily run the full compatibility tests
locally using Docker. What the script does is spin up a Docker container
for each supported Node.js version and runs the compatibility tests in
that container.

* Update Contributing Guidelines

Update the Contributing Guidelines section on compatibility testing to
include the possibility to use the newly created script for running
compatibility tests on all supported Node.js version. To achieve this,
the `"scripts"` in `package.json` were updated to provide an easier (and
robust to changes) way to run this script.

As this required some name changes to existing scripts, all references
to the pre-existing compatibility testing script were updated as well.