Instruct ESLint to report unused disable directives (#416)
Following [1], this adds the `--report-unused-disable-directives` flag
when running ESLint. This ensures that ESLint fails and reports on
useless `eslint-disable` comments.

The flag was chosen over adding it to the eslintrc file because the
former makes ESLint fail (non-zero exit code) whereas the latter
doesn't. We want to hard fail as useless ignore directives should always
be removed.

--
1. https://eslint.org/docs/latest/user-guide/configuring/rules