Add markdownlint (#425)
* Add markdownlint-cli as a dependency

Add the markdownlint-cli package as a dependency to allow for linting
MarkDown files in a convenient manner.

* Configure markdownlint

First, add an npm command to run markdownlint on the MarkDown files in
the project. Second, run this command in the git hooks and CI. Third,
add a configuration file for markdownlint to prevent it from warning
about

- Lines that are too long lines if the line belong to a table.
- Duplicate titles if they're not siblings (sub-headings with the same
  name are used at heading level 4 and lower).

* Add use of markdownlint to the Contributing Guidelines

Lastly, document the existence of this new command in the Contributing.
This is present in a new section dedicated to linting. As such, other
non-Prettier lint commands have been included in this section as well.