Narrow supported Node.js versions range to currently-out-versions (#432)
This is to prevent the expectation that this package will be compatible
with a to-be-released Node.js version. When a new Node.js version comes
out, it can be added to the list of supported versions as a new feature.

What about Node.js v11, v13, v15, and v17?
These are all intermediate release that no-one should be using after the
even release following it has come out. Any users currently using any of
these should immediately upgrade to at least the next even release.

That being said, I'm happy to patch support for any of these back into
the v1 release line on demand.