Convert Codecov validation CI check into a local script (#433)
Take the Codecov validation CI check script and turn it into a shell
script that can be easily run independently of the CI. Use it in the CI
to avoid repetition. Also add messaging to make it a little bit more
user friendly when running it locally.

Also update the `lint:sh` command to lint this script (and others in the
same directory).