Update manual release guidelines
Primarily adjust it so that it asks the releaser to create a release
branch and release Pull Request, avoiding pushes to the `main` branch.

Additionally, adjust some wording for clarity:
- The tag to be created is a **git tag**
- The git tag annotation should contain the changes **from the
  changelog**.
- The continuous delivery might still deploy to npm if you follow the
  manual process.

Also, following the addition of new steps to the process, the numbered
list now leverages MarkDown smart numbered lists by just numbering all
items with "1" (converted into 1, 2, ... when rendered).

Also add a (missing) step for creating a GitHub Release.

Lastly, also update the verification step to
1. Use `npm clean-install` for better reproducibility.
2. Run more tests, just in case.