Configure incremental mode for mutation testing
Update the Stryker configuration file to include configuration for
incremental mode. Given poor support when using `commandRunner` it is
disabled by default, but the `--incremental` flag can still be used to
enable it.

By "poor support" I primarily mean that Stryker isn't able to tell when
a test file changed and what mutants need to be rerun as a result.