Use incremental mode for mutation testing in CI (#446)
Leverage Stryker's incremental mode [1] to improve the speed of the
mutation testing check in the CI.

The speed up is substantial. Given no unit test is changed in a given
Pull Request or push, the duration of the mutation tests in CI drops
from 16m 14s to 0m 17s, or about ~1.7% of the original runtime.

Note that by default incremental mode has been disabled because of poor
support when using commandRunner (but the --incremental flag can still
be used to enable it). In particular, changes to the unit tests won't be
detected by Stryker. So, to avoid false results due to that by an
inexperienced user, it's disabled by default. Experienced users are
still free to use incremental mode during development.

--
1. https://stryker-mutator.io/blog/announcing-incremental-mode