Update Windows test fixtures for control characters
Update the Windows test fixtures to include samples of strings with the
control characters:
- `<backspace> ('\\b')`
- `<carriage return> ('\\r')`
- `<escape> ('\\u001B')`
- `<control sequence introducer> ('\\u009B')`
and how these should be escaped for each supported Windows shell. The
tests for `<end of line> ('\\n')` were also updated due to its relates
to the carriage return tests.

Note that for CMD, there's no actual changes in the expected behaviour
when it comes to <end of line> and <carriage return> because it is
already the case these will always be removed.

In most cases, these characters should be escaped because they could
allow for controlling input outside the range of the input. The most
prominent example of this is the carriage return, which resets the
cursor to the start of the line.

The only exception to this is the exact sequence "\n\r" (newline
representation on Windows), which will be preserved as it's known safe
(even if it can be considered unnecessary on Unix).