Combine similar replacements for Windows
Refactor win.js to combine replacements where the expression and
replacement string are similar into one call to `.replace()`.

This is a very minor performance improvement. In the worst case scenario
the number of steps doesn't change (all characters in the input string
are evaluated against all characters in the regular expression), but in
a better case scenario there are fewer comparisons because escaped
characters from regex $i aren't evaluated against the replacements
prior.

It also avoid any ambiguity about the order of these replacements as
well as to which replacement regex a particular character belongs