Escape control characters (#456)
* Update test fixtures

Update the Unix and Windows test fixtures to include samples of strings
with the control characters:

- `<backspace> ('\\b')`
- `<carriage return> ('\\r')`
- `<escape> ('\\u001B')`
- `<control sequence introducer> ('\\u009B')`

and how these should be escaped for each supported Unix shell. The tests
for `<end of line> ('\\n')` were also updated due to its relates to the
carriage return tests. Also, other test cases that include the carriage
return character were update according to how it should be escaped.

In most cases, these characters should be escaped because they could
allow for controller input outside the range of the input. The most
prominent example of this is the carriage return, which resets the
cursor to the start of the line.

The only exception to this is the exact sequence "\r\n" (newline
representation on Windows), which will be preserved as it's known safe
(even if it can be considered unnecessary on Unix).

* Escape abusable escape characters

Update escaping for all Unix and all Windows shells to escape the
characters listed above in all situations with the soul exception of the
carriage return character when it appears in exactly as `\r\n` for Bash,
Dash, PowerShell, and Zsh (the carriage return character is already
always stripped for CMD).

* Update eslint-plugin-regex configuration

Disable the `regexp/no-control-character` rule because this project
occasionally requires matching control characters for the purposes of
escaping or removal.

* Update argument normalization for fuzzing

Update the fuzz normalization for all shells in alignment with the
changes to the escaping of control characters in the source code.

* Update test macro `escape` for unit tests

Update the macro title for the test that runs fixutures to account for
the newly introduced control characters.

* Update CHANGELOG