Update test fixtures for whitespace (#461)
* Update test fixtures for whitespace

Improve the test fixtures for whitespace (and, relatedly, control
characters and the null character) with proper separation and naming.

* Increase timeout for integration tests

The tests are starting to run out of time in CI due to the quite
significant number of test fixtures. As an immediate solution, give the
integration tests a larger timeout in the CI to ensure they pass.