Use action artifacts to get index.cjs in compatibility test jobs
Update the compatibility test jobs to download index.cjs as a artifact
rather than building the file in the job itself. This is necessary now
because rollup v3 is not compatible with some of the older version of
Node.js supported by this project. So, by building the file on the
development Node.js version for this project and using it for testing
this problem is avoided. This also makes the setup more resilient for
the future, so even if the supported Node.js version of this project and
rollup overlap, this change shouldn't be reverted.