Update compatibility testing pre-hook
Remove the pre-hook for running the compatibility test suite as it
might fail when running on an older supported Node.js version (this is
because the transpiling doesn't have to work on the older Node.js
version). Instead, transpile prior to running `test:compat-all` as,
presumably, this is invoked from a compatible Node.js version and the
transpiled file is required for all those tests.

Update the Contributing Guidelines accordingly.