Update cmd.exe fuzz preparation
Update the arg preparation when fuzzing cmd.exe by accounting for newly
stripped characters by Shescape when preparing arguments containing
forward slashes. In particular in addition to `\0`, Shescape also strips
`\u0008`, `\u001B`, and `\u009B`.

Accordingly, fix the comment explaining to replacement as it incorrectly
claims that `\0` is ignored (while in fact it's Shescape that removes
it).