Address Windows-shell fuzz crashes (#478)
* Add two item to the fuzz corpus

Add one item that cause crashes when fuzzing cmd.exe in the
repository in the CI at [1]. Based on local testing this also affects
PowerShell (confirmed in CI).

Add another item that cause crashes when fuzzing PowerShell in
the repository in the CI at [1]. Based on local testing this does not
affect cmd.exe (confirmed by CI).

*  Update cmd.exe fuzz preparation

Update the arg preparation when fuzzing cmd.exe by accounting for newly
stripped characters by Shescape when preparing arguments containing
forward slashes. In particular in addition to `\0`, Shescape also strips
`\u0008`, `\u001B`, and `\u009B`.

*  Update PowerShell fuzz preparation

Update the arg preparation when fuzzing PowerShell by accounting for
newly stripped characters by Shescape when preparing arguments
containing forward slashes. In particular in addition to `\0`, Shescape
also strips `\u0008`, `\u001B`, and `\u009B`.

* Update fuzz preparation comments

Accordingly with the previous two changes, fix the comment
explaining to replacement as it incorrectly claims that `\0` is
ignored (while in fact it's Shescape that removes it).

* Improve PowerShell fuzz preparation

Improve how the carriage return character is considered when preparing
arguments for PowerShell. In particular, rather than only considering it
at the end of the argument, consider it any time it's not followed by a
line feed.


--
1. 6c31238de6bad3977fac5cf4844890883a615c3d