Disable auditing on npm commands
Update the npm configuration file to disable automated auditing with
`npm audit` when running (certain) npm commands (such as `npm install`).

Doing this to reduce potential noise by `npm audit` that is
intentionally suppressed by the `npm run audit` script. The automated
audit checks in the CI, as well as instructions on auditing in the
Contributing Guidelines should suffice for auditing dependencies for
this project.