Drop "audit:runtime" script
Auditing runtime dependencies only is unnecessary as this is a library,
it is generally not problematic for runtime dependencies being depended
to have known vulnerabilities as code is never run in an interesting
setting and users of the library can upgrade to a safe version because
of the use of version ranges.

Note that this assessment excludes vulnerabilities in the malware
category.