Set up auditing of latest released version (#495)
* Create release job to update major version's sliding branch

* Audit the "v1" branch nightly

The "v1" branch should always point to the same commit as the latest
v1.X.Y tag. Auditing it nightly to get alerts if it allows for depending
on a vulnerable version of the dependency. If so, the dependency should
be updated to depend on the lowest non-vulnerable version so that, in
the next release, end-users are automatically on a safe version.

* Audit only runtime dependencies for release branches