Add licensee
Add licensee [1] as a dependency to be used as a license validation
tool. An initial configuration is included that defines an allowlist of
licenses based on the (permissive) licenses used by dependencies at this
point in the commit history. Note that 3 dependencies are not covered
because they don't specify their license in a standard manner. As a
result, running the license checker will fail (i.e. exit with a non-zero
exit code).

--
1. https://www.npmjs.com/package/licensee