Use `licensee` for license scanning (#505)
* Add licensee

Add licensee [1] as a dependency to be used as a license validation
tool. An initial configuration is included that defines an allowlist of
licenses based on the (permissive) licenses used by dependencies at this
point in the commit history. Note that 3 dependencies are not covered
because they don't specify their license in a standard manner. As a
result, running the license checker will fail (i.e. exit with a non-zero
exit code).

* Run license check in the CI

Add a new job to the "Check" workflow to validate the licenses of
dependencies using the `npm run check-licenses` command.

* Manually review "NOT APPROVED" packages

Update the licensee [1] configuration to include exception for packages
that were "NOT APPROVED" yet. These were manually reviewed:

- "prelude-ls": MIT licensed, see `package.json` line 30 at commit
  d69be8fd8a682321ba24eced17caf3a1b8ca73b8 of the repository.
- "docopt": MIT licensed, see:

      cat node_modules/docopt/package.json | grep version
      cat node_modules/docopt/package.json | grep license

- "scss-parser": BSD-3-Clause licensed, see:

      cat node_modules/scss-parser/package.json | grep version
      cat node_modules/scss-parser/package.json | grep license
      cat node_modules/scss-parser/README.md
    
* Add section about license validations to the Contributing Guidelines

--
1. https://www.npmjs.com/package/licensee