Merge win property tests with win unit tests
Take the property test for the win.js module and add them to the unit
tests for the win.js module. In most cases this takes the form of
replacing an existing unit test by the "same" property test.

The exception to this is the tests for escaping and quoting for
supported shells. Here, the fixture-based tests are kept to guarantee
certain behaviour for specific example. In the case of quoting, the
property tests serve to test property that should hold for all inputs
(there are no interesting properties being tested in the pre-existing
property tests for escaping, hence it's omitted).