Merge platforms property tests with platforms unit tests
Take the property test for the platforms.js module and add them to the
unit tests for the platforms.js module. This amounts to replacing the
"environment variables are missing" unit tests by the corresponding
property test, and replacing the various `macros.platformHelpers` tests
by (new) property tests that vary the environment variables (the now
unused `macros.platformHelpers` is removed as well).

To support these new tests, the platform and OSTYPE arbitraries have
been updated to accept an optional argument to choose a category of
operating system.

Also, a miscellaneous change to win/escape.test.js: removing an unused
import.