Merge main property tests with main unit tests
Take the property test for the main.js module and add them to the unit
tests for the main.js module. By and large this amounts to replacing
unit tests (for a specific value) by their corresponding property tests.
In some cases, where the value being tested is limited to a small finite
set, the property test is adopted in such a way that there is a 100%
guarantee each possible value is tested.

Additionally, the previously shared setup functions are now re-embedded
into the unit test modules.