Merge property tests into unit tests (#519)
* Merge win property tests with win unit tests

Take the property test for the win.js module and add them to the unit
tests for the win.js module. In most cases this takes the form of
replacing an existing unit test by the "same" property test.

The exception to this is the tests for escaping and quoting for
supported shells. Here, the fixture-based tests are kept to guarantee
certain behaviour for specific example. In the case of quoting, the
property tests serve to test property that should hold for all inputs
(there are no interesting properties being tested in the pre-existing
property tests for escaping, hence it's omitted).

* Merge unix property tests with unix unit tests

Take the property test for the unix.js module and add them to the unit
tests for the unix.js module. In most cases this takes the form of
replacing an existing unit test by the "same" property test.

The exception to this is the tests for escaping and quoting for
supported shells. Here, the fixture-based tests are kept to guarantee
certain behaviour for specific example. In the case of quoting, the
property tests serve to test property that should hold for all inputs
(there are no interesting properties being tested in the pre-existing
property tests for escaping, hence it's omitted).

The property test for the default shell has been removed as the function
has no inputs in the source code.

* Merge platforms property tests with platforms unit tests

Take the property test for the platforms.js module and add them to the
unit tests for the platforms.js module. This amounts to replacing the
"environment variables are missing" unit tests by the corresponding
property test, and replacing the various `macros.platformHelpers` tests
by (newly created) property tests that vary the environment variables.

* Merge main property tests with main unit tests

Take the property test for the main.js module and add them to the unit
tests for the main.js module. By and large this amounts to replacing
unit tests (for a specific value) by their corresponding property tests.
In some cases, where the value being tested is limited to a small finite
set, the property test is adopted in such a way that there is a 100%
guarantee each possible value is tested.

Additionally, the previously shared setup functions are now re-embedded
into the unit test modules.

* Remove now-unused test macros