Remove odd call to `getEscapeFunction` from main.js
This call didn't make sense and indicates knowledge about the internal
workings of the platform-specific modules is seeping into main.js

Fix by moving the conditional into the (previously named) `getBasename`
function.

Enable skipped tests that now pass, update the tests for the platform
modules in accordance with this change.