Try Jazzer.js
Remove js-fuzz and replace it with Jazzer.js - trying it out for now.

Jazzer.js is based on LibFuzzer (whereas js-fuzz has it's own AFL-based
fuzzing engine). Jazzer.js is currently actively maintained & developed,
in contrast to js-fuzz which is not actively being developed and not
well maintained (as an example, js-fuzz brings in transitive deps with
deprecation warnings whereas Jazzer.js doesn't).

The Unix- and Windows-specific workflows have been updated to trigger on
changes to the fuzz script. This make sense more generally, but is done
in this commit specifically to test Jazzer.js in the CI setup.