Run release workflow on main if manifest has a new version number
Update the release workflow to run a job to check if a release is
necessary by looking if there's a release (by tag) for the current
version specified in the project manifest (`package.json`). If no
such tag exists, run the publish workflow. This also means that the
publish workflow now includes pushing a tag for the release. For
simplicity, release tags are no longer annotated.

Note that this removed the "verification" (linting, testing, etc.)
from the publish workflow as this is now inherent in the fact that
a Pull Request is required to release.

The Release Guidelines and Release Pull Request description have
been updated accordingly.