Pin all runner images in GitHub Actions workflows
Update all GitHub Actions workflows to use a specific runner image for
all jobs. The chosen images are based on what the documented `latest`
images are on December 9, 2022. Upgrading to more recent images may be
done seperately.

Doing this should help with reproducability. It's currently not known if
Dependabot will keep these up-to-date automatically.