Pin all runner images in GitHub Actions workflows (#576)
* Pin all runner images in GitHub Actions workflows

Update all GitHub Actions workflows to use a specific runner image for
all jobs. The chosen images are based on what the documented `latest`
images are on December 9, 2022. Upgrading to more recent images may be
done separately.

Doing this should help with reproducibility. It's currently not known if
Dependabot will keep these up-to-date automatically.

* Make OS matrixed job names constant

Update matrix jobs that vary the OS to have a constant name while the OS
version can change over time.

The primary advantage of doing this is that the job can be made required
and required jobs don't need to change if the OS version is changed.

* Rework reusable-fuzz to avoid repetition of OS in job matrix

Also update the check-*.yml workflows to run if it or the reusable fuzz
workflow changed. This makes sense and also allows for easily testing
these changes.

* Extract fuzz matrix script to .js file (& checkout repo to run script)