Create asdf `.tool-versions` file
Create a `.tool-versions` file with currently untracked dependencies
(that we want to track). This file is first a configuration format for
asdf [1], but is also human-readable and thus can be used to know and
track what versions of dependencies _should_ be used.

Using asdf with this tool-versions file requires installing plugins for
ShellCheck and actionlint.

--
1. https://asdf-vm.com/