Proof of concept of escaping spaces (` `) for Unix shells
Example implementation of escaping whitespace, in this case spaces, for
Unix shells. In particular when escaping unquoted (and "unprotected",
unlike e.g. calling `execFile` with a falsy `shell` value) arguments.

Since whitespace may change how other things are escaped, to be on the
conservative it is now escaped last. To minimize changes `\n` is escaped
after `[\t ]` (for now).

Fuzzing required some changes as a previous side effect is removed as a
result of this change. In particular, spaces (` `) introduced as a
result of replacing leading/trailing newlines would be removed due to
the trimming logic. This no longer happens due to the source code
changes in this commit. This is overcome by using an intermediate
representation that's guarenteed not to be in the string (due to line
55).

The "Hello world" sample string test case has been removed as it would
actually belong in the `"<space> (' ')"` data group. Since this kind of
string is already covered it's removed.