Update "Publish / npm"
Assign the publish workflow's npm job to the GitHub Actions env [1]
named "npm" in order to limit the scope of the `NPM_TOKEN` secret from
all workflows to workflows with the right environment.

The URL should show up in the GitHub Actions UI and is intended for the
deployer's convenience.

--
1. https://docs.github.com/en/actions/deployment/targeting-different-environments/using-environments-for-deployment