Try to update tooling nightly (#613)
Add a new job to the nightly workflow to try and update tooling (see
`.tool-versions`) to the latest available version and open a Pull
Request if there are any changes. This should help ensure tooling is
kept up-to-date.

For now, the list of tools is duplicated between the .tool-versions file
and this workflow (see the `matrix`). This may be improved upon in the
future.