Make fuzzing with default shell explicit
Remove the "WithoutShell" variants of fuzzing `check` functions in favor
of fuzzing explicitly without a `FUZZ_SHELL` set.

Doing this avoids overlap between seemingly unrelated fuzz runs, for
example: a fuzz error detected for the default shell will be found by
fuzzing using any shell. This may confusing (if the stacktrace isn't
read carefully) and leads to "incorrectly" failing CI jobs (e.g. fuzzing
PowerShell fails because of an escape bug for CMD).

An explicit example of how to fuzz with the default shell has been added
to the Contributing Guidelines, even though simply not defining
`FUZZ_SHELL` has the same effect.

The `getFuzzShell` helper function is updated to explicitly return
`undefined` if the specified FUZZ_SHELL is `""`. This is in accordance
with the changes to the Contributing Guidelines.

Note that `fork` does not support defining a shell explicitly so isn't
subject to these changes.