Update README.md (#630)
* Make "Further Reading" the last section

As "going further" kinda implies there's nothing left here.

* Shorten "Shells" section

Make the "Shells" section in the README shorter. This should help make
the README feel shorter and by extension "easier" to read, increasing
the likelihood that users read it.

* Make installation snippet copy-able

Update the snippet for installing Shescape to make it copy-able as a
whole. This makes it easier to use the snippet for the purposes of
installing, reducing the likelihood of users being affected by
typosquatting.

* Re-order README.md

Move the recipes and API sections under the "Usage" section as they're
about usage. This also makes the README flow more as the statement "Use
shescape" is now followed by sections on how to use it (rather than a
list of supported shells).