Adjust ` nick-fields/retry` config
For non-zero exit codes to cause retries, the *last* command must fail.
Thus, as the first command is expected to be flaky, convert the command
into a one-liner.

Remove `retry_on: error` to allow for retrying on timeouts, e.g. if the
network hangs.

Set `retry_wait_seconds: 30` for the purposes `timeout_minutes: 1` was
initially set. Keep `timeout_minutes: 1` in accordance with the removal
of `retry_on: error`.