Create initial metamorphic test case
Create a first "truly" metamorphic test case for the `quote` function.
This test asserts that adding an alphanumeric string to the value being
quoted is the same as adding that string afterwards (in between the
quotes, of course). This string has to be alphanumeric as otherwise it's
likely one of it's characters has to be escaped, in which case the
metamorphic relation no longer holds.