Revert 10a146b3ea1e0dde8e4b9a3d0d6151071d92534d
This property does not hold in general. This can be seen from the fact
that a similar test for `escape` does not work due to additional
characters changing how things should be escaped. Just because this
currently does not affect escaping when `quote` is used does not mean
this property should hold always.