Improve integration tests (#647)
* Update integration test assertions for `escapeAll` and `quoteAll`

In the integration tests for `escapeAll` and `quoteAll` use the fact
that output elements must equal the result of their singular counters
(`escape` and `quote`) to make stronger assertions than the type (and,
in the case of `quoteAll`, general structure).

* Correct prototype pollution integration tests for `quote` and `escape`

These functions don't typically accept arrays. Probably a copy-paste
error from their `-All` counterparts.

* Create metamorphic test cases for `escapeAll` and `quoteAll`

Test that an additional argument when escaping or quoting a list of
arguments does not have an effect on the other arguments being escaped
or quoted.