Add preliminary C shell support
Start support for the C shell (on Unix systems) with a basic
implementation that should support usage **without** interpolation
(based on preliminary results of fuzzing).

Fixture objects for C shell were initialized empty to ensure all tests
can run, but no test cases have been added as of yet.

Support for the C shell is already stated in the README.md, just so we
don't forget to add it later.