Resolve csh hanging
Fix the problem where the csh would hang on certain (escaped) arguments.
Details about the fix can be found in the comment in `src/unix.js`. The
`utf8` package [1] was added for convenient manupulation of strings.
The package may be replaced with a local implementation in the future.

Tests were updated accordingly. Test fixtures for the C shell have been
updated based on what was failing. Additionally, given the nature of
this change, a custom property test is created that tests this specific
behaviour. It constructs random strings containing random characters
with the property described in the code comment, and verifies these
characters are always escaped appropriately.

An additional fuzz corpus entry related to this bug has been added. The
previously added entry[^1] contains a character that utf-8 encodes as:

    0xDF 0xA0

The newly added entry[^2] contains a character that utf-8 encodes as:

    0xEA 0x98 0xA0

Both fail in the same way, but the `0xA0` character is in a different
position.

--
^1 1f575e02a1cd02b83266306ef195851dfd087e20708452d1dcc2c81e4ef3ccbd
^2 dc03b80d78256c11729ab7b488d3bba7be3b541196de687c42ba37b214a2ddb9

--
1. https://www.npmjs.com/package/utf8